+++ lib/oxaiworkers/state_helper.rb
@@ -3,7 +3,7 @@
 module OxAiWorkers
   module StateHelper
     def log_me(transition)
-      # OxAiWorkers.logger.debug("`#{transition.event}` was called to transition from :#{transition.from} to :#{transition.to}")
+      # OxAiWorkers.logger.info("`#{transition.event}` was called to transition from :#{transition.from} to :#{transition.to}")
     end
   end
 end